requirements.txt
requests==2.28.1
python-dotenv==0.20.0
# discord==2.0.0
discord-py-interactions==4.3.1
